// openapi/client/models/createSaleOrderDtoItemsItem.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Easy Óleo API
 * Distribuição de óleo automotivo. Os schemas gerados a partir dos DTOs são consumidos no frontend via Orval/OpenAPI.
 * OpenAPI spec version: 1.0
 */

export type CreateSaleOrderDtoItemsItem = {
  /** @minLength 1 */
  productId: string;
  /** @exclusiveMinimum 0 */
  quantity: number;
};